+++ Chord Provider/PDFBuild/SongElements/PDFBuild+LyricsSection+Line+Part.swift
@@ -16,12 +16,10 @@ extension PDFBuild.LyricsSection.Line {
     /// Display a part of a line of a lyrics section of the song
     class Part: PDFElement {
 
+        /// The chord of the part
+        let chord: NSAttributedString?
         /// The text of the part
-        /// - Note: A merge of the optional chord and optional lyrics part
-        let text = NSMutableAttributedString()
-        /// The size of the part
-        /// - Note: Used by the `Line` element to define the rectangle
-        let size: CGSize
+        let text: NSAttributedString
         /// All the chords of the ``Song``
         let chords: [ChordDefinition]
 
@@ -31,34 +29,28 @@ extension PDFBuild.LyricsSection.Line {
         ///   - chords: All the chords of the song
         ///   - settings: The application settings
         init(part: Song.Section.Line.Part, chords: [ChordDefinition], settings: AppSettings) {
+
+            var fontOptions = settings.style.fonts.text
+            fontOptions.color = settings.style.theme.foreground
+
             self.chords = chords
-            if chords.isEmpty {
-                self.text.append(
-                    NSAttributedString(
-                        string: "\(part.text)",
-                        attributes: .attributes(.text, settings: settings)
-                    )
-                )
-            } else {
+
+            let lyrics = "\(part.text)\u{200c}"
+            let text = lyrics.toMarkdown(fontOptions: fontOptions, scale: settings.pdf.scale)
+            self.text = NSAttributedString(text)
+
+            if !chords.isEmpty {
                 var chordString: String = " "
                 if let chord = chords.first(where: { $0.id == part.chord }) {
                     chordString = chord.display
                 }
-                self.text.append(
-                    NSAttributedString(
-                        /// Add a space behind the chord-name so two chords will never 'stick' together
-                        string: "\(chordString) \n",
-                        attributes: .partChord(settings: settings)
-                    )
+                self.chord = NSAttributedString(
+                    string: "\(chordString)",
+                    attributes: .attributes(.chord, settings: settings)
                 )
-                self.text.append(
-                    NSAttributedString(
-                        string: "\(part.text)",
-                        attributes: .attributes(.text, settings: settings, scale: settings.pdf.scale)
-                    )
-                )
+            } else {
+                self.chord = nil
             }
-            self.size = text.size()
         }
 
         /// Draw the **part** element
@@ -67,27 +59,38 @@ extension PDFBuild.LyricsSection.Line {
         ///   - calculationOnly: Bool if only the Bounding Rect should be calculated
         ///   - pageRect: The page size of the PDF document
         func draw(rect: inout CGRect, calculationOnly: Bool, pageRect: CGRect) {
-            let textBounds = text.boundingRect(with: rect.size, options: .usesLineFragmentOrigin)
-            if !calculationOnly {
-                text.draw(with: rect, options: textDrawingOptions, context: nil)
+            /// The width of the part
+            var width: Double = 0
+            /// The size of the lyrics
+            let textBounds = text.boundingRect(with: rect.size)
+            if let chord {
+                let textBounds = chord.boundingRect(with: rect.size)
+                if !calculationOnly {
+                    chord.draw(with: rect, options: textDrawingOptions, context: nil)
+                }
+                let height = textBounds.height + 2 * textPadding
+                width = textBounds.width
+                rect.origin.y += height
+                rect.size.height -= height
+            } else {
+                /// I don't know why...
+                rect.origin.y += textBounds.height * 0.5
+                rect.size.height -= textBounds.height * 0.5
             }
-            let height = textBounds.height + 2 * textPadding
+
+            rect.origin.y += textBounds.height * 0.5
+            rect.size.height -= textBounds.height * 0.5
+            if !calculationOnly {
+                text.draw(with: rect)
+            }
+            let height = (textBounds.height * 0.5)
+            /// Remember the biggest width
+            width = max(width, textBounds.width)
+            /// Set the rect
             rect.origin.y += height
             rect.size.height -= height
+            rect.origin.x += width
+            rect.size.width -= width
         }
-    }
-}
-
-extension PDFStringAttribute {
-
-    // MARK: Part chord styling
-
-    /// Style attributes for the chord of the part
-    static func partChord(settings: AppSettings) -> PDFStringAttribute {
-        let font = settings.style.fonts.chord.nsFont(scale: settings.pdf.scale)
-        return [
-            .foregroundColor: settings.style.fonts.chord.color.nsColor,
-            .font: font
-        ]
     }
 }