+++ Chord Provider/PDFBuild/SongElements/PDFBuild+LyricsSection+Line.swift
@@ -17,8 +17,6 @@ extension PDFBuild.LyricsSection {
     class Line: PDFElement {
         /// The parts in the line
         let parts: [Part]
-        /// The height of a single part
-        let height: CGFloat
 
         /// Init the **line** element
         /// - Parameters:
@@ -31,7 +29,6 @@ extension PDFBuild.LyricsSection {
                 items.append(Part(part: part, chords: chords, settings: settings))
             }
             self.parts = items
-            self.height = items.first?.size.height ?? 0
         }
 
         /// Draw the **line** element
@@ -40,23 +37,19 @@ extension PDFBuild.LyricsSection {
         ///   - calculationOnly: Bool if only the Bounding Rect should be calculated
         ///   - pageRect: The page size of the PDF document
         func draw(rect: inout CGRect, calculationOnly: Bool, pageRect: CGRect) {
-            var partOffset: CGFloat = 0
+            var maxHeight: CGFloat = 0
+            var cellRect = rect
             for part in parts {
-                var cellRect = CGRect(
-                    x: rect.minX + partOffset,
-                    y: rect.minY,
-                    width: part.size.width,
-                    height: part.size.height
-                )
-                partOffset += part.size.width
-                /// Add a bit extra width if the text ends with a `space` and there are chords
-                if !part.chords.isEmpty, part.text.string.last == " " {
-                    partOffset += 2 * textPadding
-                }
+
                 part.draw(rect: &cellRect, calculationOnly: calculationOnly, pageRect: pageRect)
+
+                maxHeight = max(maxHeight, rect.height - cellRect.height)
+
+                cellRect.origin.y = rect.origin.y
+                cellRect.size.height = rect.size.height
             }
-            rect.origin.y += height + 2 * textPadding
-            rect.size.height -= height + 2 * textPadding
+            rect.origin.y += maxHeight
+            rect.size.height -= maxHeight
         }
     }
 }